fix:split TestimonialCard
